added client to render dynamic route separately
